Read the certificate/private key from keystore.
Reimplement the retrievel of the cert/key data from keystore
and use "blob://" to indicate the data is a blob, this will
save us to touch the certificate handling part in tls_openssl.c.
However, since the original cert handling is not complete for
dealing with blob data. We have to convert the blob data to DER
format if it is a PEM one.